fix(trusty-mpm): fix CI lint failures on PR #3565 (500-line cap + doc-pointer)
Two CI gates in this repo weren't covered by the fmt/clippy/test gate used
while developing the #3556 fix, so they surfaced only after the PR opened:

- 500-line file-size cap: `core::deploy_validate::mod.rs` grew to 549 SLOC
  once the AgentFrontmatterInvalid probe and its two regression tests
  landed. The cap counts by filename pattern, not by `#[cfg(test)]`, so an
  inline `mod tests { ... }` block counts against the 500 SLOC production
  cap. Split the test module out to `deploy_validate/tests.rs` (loaded via
  `#[path = "tests.rs"] mod tests;`), mirroring the existing
  builder.rs/builder_tests.rs and deployer.rs/deployer_tests.rs precedent in
  trusty-agents-common. Pure extraction — no logic changes, all 23 tests
  unchanged and passing.

- Doc-comment pointer lint: `render_scalar`'s doc comment in builder.rs cited
  a test named `compose_description_with_colon_is_quoted`, but the actual
  test (added in the #3556 commit) is named
  `compose_description_with_colon_is_quoted_and_strict_yaml_valid`. Fixed
  the pointer to match.

No functional changes to the frontmatter fix or the deploy-time validation
added in the prior commit.
